sales_platform_auth_project/requirements.txt
django
djangorestframework
djangorestframework-simplejwt
psycopg2-binary
gunicorn

dj-rest-auth[with_social]
django-allauth
requests-oauthlib
cryptography

python-dotenv
django-cors-headers
